+++ packages/clerk-js/src/ui/contexts/ClerkUIComponentsContext.tsx
@@ -417,6 +417,7 @@ export const useCreateOrganizationContext = () => {
 
   return {
     ...ctx,
+    skipInvitationScreen: ctx.skipInvitationScreen || false,
     navigateAfterCreateOrganization,
     componentName,
   };
